:truck: refactor: move models, DTO, errors to DataAccess layer
