+++ server/DataAccess/AppException.cs
@@ -1,6 +1,6 @@
 using System.Globalization;
 
-namespace server.Helpers;
+namespace server.DataAccess;
 
 // custom exception class for throwing application specific exceptions (e.g. for validation) 
 // that can be caught and handled within the application
